Move flag passing
